requirements.txt
fastapi
uvicorn[standard]
openai
opencv-python
requests
websockets
python-dotenv
telethon
